// src/store/modules/home/types.ts
export enum HomeActionType {
  GetHomeDatas = 'getHomeDatas',
  RecordForm = 'recordForm',
  GetMyWorkflow = 'getMyWorkflow',
}

export enum HomeMutationType {
  HandleHomeDatas = 'handleHomeDatas',
  SetLocalKey = 'setLocalKey',
  PushRecentFormCode = 'pushRecentFormCode',
  GetLocalCardSetting = 'getLocalCardSetting',
  SetMyWorkflow = 'setMyWorkflow',
}